Move NotificationsApi
