+++ app/lib/api/apis/constants.js
@@ -4,3 +4,4 @@ export const TOKEN_ENDPOINT = '/token';
 export const USERS_ENDPOINT = '/users';
 export const PASSWORD_ENDPOINT = '/password';
 export const CONFIRMATION_ENDPOINT = '/confirmation';
+export const NOTIFICATIONS_ENDPOINT = '/notifications';
