+++ app/lib/api/apis/index.js
@@ -1,7 +1,9 @@
 // @flow
 
 import AuthenticationApi from './authentication';
+import NotificationsApi from './notifications';
 
 export {
   AuthenticationApi,
+  NotificationsApi,
 };
