+++ app/lib/api/index.js
@@ -2,8 +2,10 @@
 
 import {
   AuthenticationApi,
+  NotificationsApi,
 } from './apis';
 
 export {
   AuthenticationApi,
+  NotificationsApi,
 };
