+++ app/modules/feed/constants.js
@@ -1,5 +1,3 @@
 // @flow
 
 export const GRAVATAR_SIZE = 64;
-
-export const ENDPOINT = '/notifications';
